requirements.txt
azure-ai-projects>=1.0.0b1
azure-ai-agents>=1.0.0
azure-identity>=1.15.0
pydantic-settings>=2.0.0
